src/main/java/ru/job4j/tracker/FindByNameAction.java
package ru.job4j.tracker;

import java.util.List;

public class FindByNameAction implements UserAction {
    private final Output out;

    public FindByNameAction(Output out) {
        this.out = out;
    }

    @Override
    public String name() {
        return "Find by name";
    }

    @Override
    public boolean execute(Input input, Store store) {
        out.println("=== Find by name ===");
        String name = input.askStr("Enter name ");
        List<Item> items = store.findByName(name);
        if (items.size() > 0) {
            for (Item item : items) {
                out.println(item);
            }
        } else {
            out.println("Item with name " + name + " not found.");
        }
        return true;
    }
}
